Rename to RenderHelpers to be more specific about usage
